tool_management, sources refactoring
