+++ map_app/__init__.py
@@ -1,11 +1,8 @@
 from flask import Flask
-from .web.static import static_bp
-from .web.api import api_bp
+from .endpoints.static import static_bp
+from .endpoints.api import api_bp
 
 def create_app():
-    """
-    Create Flask app
-    """
     app = Flask(__name__)
 
     app.register_blueprint(api_bp)
