+++ map_app/endpoints/static.py
@@ -3,7 +3,7 @@ from flask import render_template, Blueprint
 from flask import session as flask_session
 from flask import send_from_directory
 from flask import current_app
-import map_app.tools.tool_management as tool_management
+from map_app import sources
 
 static_bp = Blueprint('static', __name__)
 
@@ -18,9 +18,11 @@ def wifi_pass_map():
 
 @static_bp.route('/tools')
 def tools():
-    tool_data = tool_management.tool_list()
+    tool_data = sources.tool_list()
     return render_template('tools.html', tools=tool_data)
 
+
+# --------------GENERIC --------------------------
 @static_bp.route('/images/<path:filename>')
 def serve_image(filename):
     return send_from_directory(os.path.join(current_app.root_path, 'map_app/static'), filename)
